requirements.txt
deta==1.1.0
pandas==1.3.5
pymongo[srv]==4.1.1
python-dotenv==0.20.0
streamlit==1.10.0
streamlit_authenticator==0.1.5
XlsxWriter